Update SkipAds.user.js
Changed mid-video ad behavior. Now sets the ad's current time to the length of the video, essentially fast forwarding the ad and skipping it.

Added another debug button. At the moment, this one just attempts to do the new mid-video behavior on whatever element you type in. Basically a skip button for any video. Not useful to users but I may update it for myself for testing things.